1.2 before state persons

+++ app/src/main/java/ru/nifontbus/contactsevents/domain/data/Person.kt
@@ -2,8 +2,8 @@ package ru.nifontbus.contactsevents.domain.data
 
 data class Person(
     val displayName: String = "",
-    val groups: List<Int> = emptyList(),
+    val groups: List<Long> = emptyList(),
     val hasPhoneNumber: Boolean = false,
-    var id: Int = -1,
+    var id: Long = -1,
 //    val key: String = "",
 )
